Attempt to finish failed package removals the next time the solution is opened.
If a package directory remains in the file system after an attempted
uninstall (possibly due to loaded assemblies like in the case of
T4Scaffolding), insert an empty .deleteme file alongside the
package directory. The next time the solution is opened, NuGet will try
to clean up all the directories in the local package repository that
are marked by a .deleteme file when, in theory, the once-locked
files are no longer locked.

This affects Visual Studio only and notifies the user to restart.
--------------------------------------------------------------------------------
Made changes suggested during the first code review:

http://nuget.codeplex.com/SourceControl/network/forks/shalter/TempTools/contribution/3242

Stopped marking package directories with added or modified files for deletion.
This change requires PackageManager.ExecuteUninstall to leave the nupkg in place
so the DeleteOnRestartManager can compare the package directory contents to the
the contents of the original package contents.

Since VSPackageManager.OnUninstalled is passed an UnzippedPackage,
MarkPackageDirectoryForDeltetion requires a Func that can create a ZipPackage
from a path. This allows the method to read the nupkg contents. This requirement
might change when Issue 2595 is resolved:

http://nuget.codeplex.com/workitem/2595

+++ src/Core/PackageManager.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Globalization;
+using System.IO;
 using System.Linq;
 using System.Runtime.Versioning;
 using NuGet.Resources;
@@ -271,6 +272,7 @@ namespace NuGet
 
         protected virtual void ExecuteUninstall(IPackage package)
         {
+            string packageDirectory = PathResolver.GetPackageDirectory(package);
             PackageOperationEventArgs args = CreateOperation(package);
             OnUninstalling(args);
 
@@ -280,8 +282,15 @@ namespace NuGet
             }
 
             OnRemoveFiles(args);
-            // Remove package to the repository
-            LocalRepository.RemovePackage(package);
+
+            // Only remove nupkg and nuspec if the package directory is otherwise empty.
+            // The DeleteOnRestartManager will use the nupkg to determine whether files are left in the package directory
+            // directory because files were added or modified as opposed to locked.
+            if (!FileSystem.GetDirectoriesSafe(packageDirectory).Any() &&
+                FileSystem.GetFilesSafe(packageDirectory).All(f => PackageUtility.IsPackageFile(f) || PackageUtility.IsManifest(f)))
+            {
+                LocalRepository.RemovePackage(package);
+            }
 
             Logger.Log(MessageLevel.Info, NuGetResources.Log_SuccessfullyUninstalledPackage, package.GetFullName());
